refactor(act): rename restore-validate.ts to restore.ts
Single-word file names — Act convention. `restore-validate.ts` was
a holdover from when the file only held the validator; it grew the
iteration helper too and the compound name stopped fitting.

Sits alongside the other public utility modules in `libs/act/src/`:
`ports.ts`, `utils.ts`, `config.ts`, `act.ts`, `signals.ts`. The
file holds both audiences' helpers — `validateRestoreRow` for app
code pre-flighting backups, `runRestore` for adapter implementers
driving the wipe-and-rebuild loop. Same audience-mix as `ports.ts`,
where `store()`/`cache()`/`log()` are app-facing and
`SNAP_EVENT`/`TOMBSTONE_EVENT` are adapter-facing.

`Store.restore` stays as the method name — "clean and commit" is
the right framing for what the port actually does.

Pure rename: imports updated in `index.ts`, the InMemory adapter,
and the unit spec. No semantic change. Workspace coverage stays at
100%.
